pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "omnibase-core"
version = "0.1.0"
description = "ONEX Core Framework - Base classes and essential implementations"
authors = ["OmniNode Team <team@omninode.ai>"]
license = "MIT"
readme = "README.md"
packages = [{include = "omnibase_core", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
omnibase-spi = {git = "ssh://git@github.com/OmniNode-ai/omnibase-spi.git", tag = "v0.0.2"}
pydantic = "^2.11.7"
llama-index = "^0.12.0"
fastapi = "^0.115.0"
uvicorn = "^0.32.0"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-asyncio = "^0.25.0"
mypy = "^1.13.0"
black = "^24.10.0"
isort = "^5.13.0"
ruff = "^0.8.0"

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "ISC", "ICN", "G", "INP", "PIE", "T20", "PT", "Q", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "NPY", "RUF"]
ignore = ["S101"]  # Allow assert statements in tests

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"